Add files via upload
Home automation using embedded systems has gained increasing popularity in 
recent years, as it provides users with a more convenient and efficient way of managing 
their homes. Embedded systems, which are small computers embedded in devices, are 
used to control various aspects of the home such as lighting, heating, security, and 
entertainment systems. In this paper we present an abstract for home automation using 
embedded systems, focusing on the system architecture, hardware and software 
components, and their interactions. The abstract describes the use of sensors for data 
collection and control, and microcontrollers for processing and decision-making. 
The project is implemented with the help of NodeMCU which have an inbuilt Wi-Fi module 
and the home appliances are connected to 4 channel relay modules. Wireless connection 
is created for communication between the NodeMCU and the Android application. By this 
way we can operate the home appliances like fans, lights and AC etc. The benefits of 
using embedded systems for home automation, including energy savings, increased 
security, and improved quality of life for users.